Update baseURL and environment variable name
The baseURL in the axios instance was switched to use an environment variable called "VITE_BACKEND_URL" instead of the previous value. The environment variable in the .env file was accordingly renamed from "VITE_APP_BASE_URL" to "VITE_BACKEND_URL". Additionally, the Dockerfile was updated to accept and set this new environment variable.